fix(Reorg file tree): Creating fixture folder / renaming test file
